move operators to operator directory
