+++ src/observer/sql/executor/execute_stage.cpp
@@ -26,10 +26,10 @@ See the Mulan PSL v2 for more details. */
 #include "event/sql_event.h"
 #include "event/session_event.h"
 #include "sql/executor/tuple.h"
-#include "sql/executor/table_scan_operator.h"
-#include "sql/executor/predicate_operator.h"
-#include "sql/executor/delete_operator.h"
-#include "sql/executor/project_operator.h"
+#include "sql/operator/table_scan_operator.h"
+#include "sql/operator/predicate_operator.h"
+#include "sql/operator/delete_operator.h"
+#include "sql/operator/project_operator.h"
 #include "sql/stmt/stmt.h"
 #include "sql/stmt/select_stmt.h"
 #include "sql/stmt/update_stmt.h"
