--- src/observer/sql/executor/execution_node.cpp.bak
@@ -1,54 +1,0 @@
-/* Copyright (c) 2021 Xie Meiyi(xiemeiyi@hust.edu.cn) and OceanBase and/or its affiliates. All rights reserved.
-miniob is licensed under Mulan PSL v2.
-You can use this software according to the terms and conditions of the Mulan PSL v2.
-You may obtain a copy of Mulan PSL v2 at:
-         http://license.coscl.org.cn/MulanPSL2
-THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
-EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
-MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
-See the Mulan PSL v2 for more details. */
-
-//
-// Created by Meiyi & Wangyunlai on 2021/5/14.
-//
-
-#include "sql/executor/execution_node.h"
-#include "storage/common/table.h"
-#include "common/log/log.h"
-
-SelectExeNode::SelectExeNode() : table_(nullptr)
-{}
-
-SelectExeNode::~SelectExeNode()
-{
-  for (DefaultConditionFilter *&filter : condition_filters_) {
-    delete filter;
-  }
-  condition_filters_.clear();
-}
-
-RC SelectExeNode::init(
-    Trx *trx, Table *table, TupleSchema &&tuple_schema, std::vector<DefaultConditionFilter *> &&condition_filters)
-{
-  trx_ = trx;
-  table_ = table;
-  tuple_schema_ = tuple_schema;
-  condition_filters_ = std::move(condition_filters);
-  return RC::SUCCESS;
-}
-
-void record_reader(const char *data, void *context)
-{
-  TupleRecordConverter *converter = (TupleRecordConverter *)context;
-  converter->add_record(data);
-}
-RC SelectExeNode::execute(TupleSet &tuple_set)
-{
-  CompositeConditionFilter condition_filter;
-  condition_filter.init((const ConditionFilter **)condition_filters_.data(), condition_filters_.size());
-
-  tuple_set.clear();
-  tuple_set.set_schema(tuple_schema_);
-  TupleRecordConverter converter(table_, tuple_set);
-  return table_->scan_record(trx_, &condition_filter, -1, (void *)&converter, record_reader);
-}
