+++ src/observer/sql/operator/insert_operator.h
@@ -9,36 +9,30 @@ MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
 See the Mulan PSL v2 for more details. */
 
 //
-// Created by WangYunlai on 2022/07/01.
+// Created by WangYunlai on 2021/6/7.
 //
 
 #pragma once
 
-#include "sql/executor/operator.h"
+#include "sql/operator/operator.h"
+#include "sql/parser/parse.h"
 #include "rc.h"
 
-class ProjectOperator : public Operator
+class InsertStmt;
+
+class InsertOperator : public Operator
 {
 public:
-  ProjectOperator()
+  InsertOperator(InsertStmt *insert_stmt)
+    : insert_stmt_(insert_stmt)
   {}
 
-  virtual ~ProjectOperator() = default;
-
-  void add_projection(const Table *table, const FieldMeta *field); // TODO how to handle the memory in tupleCellSpec?
+  virtual ~InsertOperator() = default;
 
   RC open() override;
   RC next() override;
   RC close() override;
 
-  int tuple_cell_num() const override
-  {
-    return tuple_.cell_num();
-  }
-
-  RC tuple_cell_spec_at(int index, TupleCellSpec &spec) const override;
-
-  Tuple * current_tuple() override;
 private:
-  ProjectTuple tuple_;
+  InsertStmt *insert_stmt_ = nullptr;
 };
