+++ src/observer/storage/common/record.cpp
@@ -2,15 +2,3 @@
 #include "storage/common/field.h"
 #include "common/log/log.h"
 #include "rc.h"
-
-RC Record::set_field_value(const Value &value, int index)
-{
-  // TODO
-  return RC::UNIMPLENMENT;
-}
-
-RC Record::set_field_values(const Value *values, int value_num, int start_index)
-{
-  // TODO
-  return RC::UNIMPLENMENT;
-}
